Prepare for testing. Move productshredder stuff into productshredder package
